demos: restore trails, move v9.7.0 release demo to trails-r3f-9.7.0
demos/trails goes back to its pristine upstream state (including its
fiber 9.6.1 pin). The release demo now lives as its own workspace
package @demo/trails-r3f-9.7.0 with updated pmndrs.json metadata,
page title and README links. Thumbnail is inherited from trails for
now.
