+++ demos/trails-r3f-9.7.0/README.md
@@ -1,0 +1,9 @@
+[![Static](https://img.shields.io/badge/demo-%23646CFF.svg?logo=html5&logoColor=white)](https://pmndrs.github.io/examples/trails-r3f-9.7.0)
+[![CodeSandbox](https://img.shields.io/badge/codesandbox-040404?logo=codesandbox&logoColor=DBDBDB)](https://codesandbox.io/s/github/pmndrs/examples/tree/main/demos/trails-r3f-9.7.0)
+[![Stackblitz](https://img.shields.io/badge/stackblitz-fff?logo=Stackblitz&logoColor=1389FD)](https://stackblitz.com/github/pmndrs/examples/tree/main/demos/trails-r3f-9.7.0)
+
+```sh
+$ npx degit pmndrs/examples/demos/trails-r3f-9.7.0
+```
+
+![](thumbnail.webp)
