+++ demos/trails-r3f-9.7.0/src/vite-env.d.ts
@@ -1,0 +1,1 @@
+/// <reference types="vite/client" />
