+++ demos/trails/src/App.jsx
@@ -1,145 +1,47 @@
-import { useEffect, useMemo, useRef, useState } from 'react'
+import { useRef } from 'react'
 import * as THREE from 'three'
 import { Canvas, useThree, useFrame } from '@react-three/fiber'
 import { useGLTF, Edges, Trail } from '@react-three/drei'
 import { Physics, useCompoundBody } from '@react-three/cannon'
 import { LayerMaterial, Depth, Fresnel } from 'lamina'
-import { FontLoader, TextGeometry } from 'three-stdlib'
 
-import boldFont from 'three/examples/fonts/helvetiker_bold.typeface.json'
 import pmndrsModel from './pmndrs.glb?url'
 import cursorModel from './cursor.glb?url'
 
-// One centered, extruded "9.7.0" shared by every block in the scene
-const font = new FontLoader().parse(boldFont)
-const text = new TextGeometry('9.7.0', { font, size: 0.7, height: 0.35, curveSegments: 16, letterSpacing: 0.02 })
-text.center()
-
 const vec = new THREE.Vector3()
-const dir = new THREE.Vector3()
-// The swarm gathers slightly behind the hero so it frames it instead of covering it
-const attractor = new THREE.Vector3(0, 0, -1)
-// Live registry of swarm bodies so the cursor can blast them apart on click
-const blocks = new Set()
-// Set by <Bursts>; the cursor calls it on click to spawn logo shrapnel at its position
-let burst = () => {}
 const white = new THREE.MeshBasicMaterial({ color: '#fefefe', toneMapped: false })
-const pink = new THREE.MeshBasicMaterial({ color: '#ff0080', toneMapped: false })
-// Accent blocks cycle through the cursor-gradient palette
-const accents = ['#ff0080', '#f7b955', '#5786f5'].map((color) => new THREE.MeshBasicMaterial({ color, toneMapped: false }))
+const black = new THREE.MeshBasicMaterial({ color: 'black', toneMapped: false })
 
-export const App = ({ amount = 11 }) => (
+export const App = ({ amount = 12 }) => (
   <Canvas dpr={[1, 2]} camera={{ position: [0, 0, 5], fov: 50 }}>
     <Physics gravity={[0, 1, 0]}>
-      <Hero />
       {Array.from({ length: amount }, (_, i) => (
-        <Version
-          key={i}
-          material={i % 4 === 1 ? accents[Math.floor(i / 4) % accents.length] : white}
-          scale={0.5 + (i % 3) * 0.25}
-          angularDamping={0.4}
-          linearDamping={0.8}
-          position={[(Math.random() - 0.5) * 4, (Math.random() - 0.5) * 4, (Math.random() - 0.5) * 2 - 1]}
-          rotation={[Math.random() * Math.PI, Math.random() * Math.PI, Math.random() * Math.PI]}
-        />
+        <Pmndrs key={i} mass={4} angularDamping={0.4} linearDamping={0.8} position={[Math.random(), Math.random(), Math.random()]} />
       ))}
       <Cursor mass={15} angularDamping={0.5} linearDamping={0.5} position={[0, 0, 10]} />
     </Physics>
-    <Bursts />
   </Canvas>
 )
 
-function Bursts() {
+function Pmndrs(props) {
   const { nodes } = useGLTF(pmndrsModel)
-  // The glb logo pivot is off-center; recenter a clone so shrapnel spins around itself
-  const logo = useMemo(() => {
-    const geometry = nodes.logo.geometry.clone()
-    geometry.center()
-    return geometry
-  }, [nodes])
-  const [bursts, setBursts] = useState([])
-  const id = useRef(0)
-  useEffect(() => {
-    burst = (origin) => setBursts((all) => [...all, { id: ++id.current, origin }])
-    return () => (burst = () => {})
-  }, [])
-  return bursts.map(({ id, origin }) => (
-    <Burst key={id} geometry={logo} origin={origin} onDone={() => setBursts((all) => all.filter((b) => b.id !== id))} />
-  ))
-}
-
-// Purely visual logo shrapnel: flies outward, tumbles, shrinks away — no physics cost
-function Burst({ geometry, origin, onDone, count = 16, life = 1.5 }) {
-  const group = useRef()
-  const age = useRef(0)
-  const particles = useMemo(
-    () =>
-      Array.from({ length: count }, (_, i) => ({
-        // random sphere, biased towards the camera so debris whooshes past the viewer
-        dir: new THREE.Vector3().randomDirection().add(vec.set(0, 0, 0.5)).normalize(),
-        speed: 5 + Math.random() * 7,
-        rotation: [Math.random() * Math.PI * 2, Math.random() * Math.PI * 2, Math.random() * Math.PI * 2],
-        spin: [(Math.random() - 0.5) * 12, (Math.random() - 0.5) * 12, (Math.random() - 0.5) * 12],
-        size: 0.4 + Math.random() * 0.8,
-        material: Math.random() < 0.45 ? accents[i % accents.length] : white
-      })),
-    [count]
-  )
-  useFrame((_, delta) => {
-    age.current += delta
-    const k = age.current / life
-    if (k >= 1) return onDone()
-    // pop in fast, decelerate, shrink to nothing
-    const envelope = Math.min(k * 10, 1) * (1 - k)
-    group.current.children.forEach((child, i) => {
-      const p = particles[i]
-      child.position.addScaledVector(p.dir, p.speed * (1 - 0.7 * k) * delta)
-      child.rotation.x += p.spin[0] * delta
-      child.rotation.y += p.spin[1] * delta
-      child.rotation.z += p.spin[2] * delta
-      child.scale.set(0.188, 0.188, 0.97).multiplyScalar(p.size * envelope)
-    })
-  })
+  const [ref, api] = useCompoundBody(() => ({
+    ...props,
+    shapes: [
+      { type: 'Box', args: [0.65, 0.65, 0.5], position: [0.18, 0.18, 0] },
+      { type: 'Box', args: [0.3, 0.3, 0.5], position: [-0.35, 0, 0] },
+      { type: 'Box', args: [0.3, 0.3, 0.5], position: [0, -0.35, 0] }
+    ]
+  }))
+  // Pull every logo back towards the center of the scene
+  useFrame(() => api.applyForce(vec.setFromMatrixPosition(ref.current.matrix).normalize().multiplyScalar(-40).toArray(), [0, 0, 0]))
   return (
-    <group ref={group} position={origin}>
-      {particles.map((p, i) => (
-        <mesh key={i} geometry={geometry} material={p.material} rotation={p.rotation} scale={0.001}>
-          <Edges threshold={20} color="black" lineWidth={1} />
-        </mesh>
-      ))}
-    </group>
-  )
-}
-
-function Block({ scale = 1, material = white, edge = 'black', ...props }) {
-  return (
-    <group {...props}>
-      <mesh scale={scale} geometry={text} material={material}>
-        <Edges scale={1.002} threshold={20} color={edge} lineWidth={1.5} />
+    <group ref={ref}>
+      <mesh scale={[0.188, 0.188, 0.97]} position={[-0.02, -0.5, 0.022]} geometry={nodes.logo.geometry} material={white}>
+        <Edges scale={1.005} material={black} />
       </mesh>
     </group>
   )
-}
-
-// One box approximating the extruded "9.7.0" word, baked to the visual scale
-const wordBox = (scale) => [{ type: 'Box', args: [2.2 * scale, 0.55 * scale, 0.35 * scale], position: [0, 0, 0] }]
-
-// Immovable billboard version number the swarm and cursor bounce off
-function Hero({ scale = 1.2 }) {
-  const [ref] = useCompoundBody(() => ({ mass: 0, position: [0, 0, 1.5], shapes: wordBox(scale) }))
-  return <Block ref={ref} scale={scale} material={pink} />
-}
-
-function Version({ scale = 1, material = white, ...props }) {
-  const [ref, api] = useCompoundBody(() => ({ mass: 4 * scale, ...props, shapes: wordBox(scale) }))
-  useEffect(() => {
-    const entry = { ref, api, scale }
-    blocks.add(entry)
-    return () => blocks.delete(entry)
-  }, [ref, api, scale])
-  // Pull every block back towards the gather point behind the hero
-  useFrame(() => api.applyForce(vec.setFromMatrixPosition(ref.current.matrix).sub(attractor).normalize().multiplyScalar(-40 * scale).toArray(), [0, 0, 0]))
-  return <Block ref={ref} scale={scale} material={material} />
 }
 
 function Cursor({ speed = 10, gradient = 0.7, ...props }) {
@@ -156,25 +58,8 @@ function Cursor({ speed = 10, gradient = 0.7, ...props }) {
   }))
   useFrame((state) => {
     vec.setFromMatrixPosition(ref.current.matrix)
-    // Chase the pointer, diving to the swarm's gather depth so it plows through the clump
-    api.velocity.set(((state.mouse.x * viewport.width) / 2 - vec.x) * speed, ((state.mouse.y * viewport.height) / 2 - vec.y) * speed, (attractor.z - vec.z) * 2)
+    api.velocity.set(((state.mouse.x * viewport.width) / 2 - vec.x) * speed, ((state.mouse.y * viewport.height) / 2 - vec.y) * speed, -vec.z)
   })
-  useEffect(() => {
-    // Click detonates: every block gets a distance-falloff impulse away from the cursor,
-    // applied slightly off-center so they tumble. The attractor then reels them back in.
-    const explode = () => {
-      vec.setFromMatrixPosition(ref.current.matrix)
-      burst(vec.toArray())
-      for (const block of blocks) {
-        dir.setFromMatrixPosition(block.ref.current.matrix).sub(vec)
-        const falloff = Math.max(dir.length(), 1)
-        dir.normalize().multiplyScalar((42 * block.scale) / falloff)
-        block.api.applyImpulse(dir.toArray(), [(Math.random() - 0.5) * 0.4, (Math.random() - 0.5) * 0.4, (Math.random() - 0.5) * 0.4])
-      }
-    }
-    window.addEventListener('pointerdown', explode)
-    return () => window.removeEventListener('pointerdown', explode)
-  }, [])
   return (
     <>
       <Trail color="red" target={trail} length={4} width={0.5} attenuation={(w) => w * 2} />
